requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
psycopg2-binary>=2.9.9
boto3>=1.28.62
mcp-server>=0.1.0